Currently, the entire KV cache is allocated as a single tensor for all the layers. As a consequence, the KV cache is either fully on the CPU, or fully offloaded to the GPU.

With this change, the KV cache is allocated on a different tensor per layer. The result is more granular control over the parts of the KV offloaded to the GPU.

In this demo, when partially offloading a model, the KV cache corresponding to the offloaded layers is also offloaded. This increases performance at the expense of more VRAM.

Is it worth it compared to just offloading more layers? I am not sure, but probably wouldn't hurt to have more flexibility.

Note: only implemented for llama models. CUDA only.

Edit: removed a few unnecessary copies that caused performance to degrade.

![6fb9afe1-eca9-480a-a62c-ed01d02503c7](https://github.com/ggerganov/llama.cpp/assets/2141330/61358346-fec6-444e-a6eb-daec65f87c03)
![83343429-82e1-4bb5-b034-8667cf242d71](https://github.com/ggerganov/llama.cpp/assets/2141330/cbc23275-9bbf-4aab-b6af-e5a2cb812348)

Llama2 70B on a single 24 GB GPU:
Device 0: NVIDIA GeForce RTX 3090 Ti, compute capability 8.6
| model                    |       size |     params | ngl | test       |       master t/s |          PR t/s |    speedup |
| ------------------------ | ---------: | ---------: | --: | ---------- | ---------------: | --------------: | ---------: |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  60 | pp 512     |     83.10 ± 0.26 |   115.29 ± 0.33 |       1.39 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  60 | tg 128     |      4.21 ± 0.05 |     4.78 ± 0.05 |       1.14 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  61 | pp 512     |     84.29 ± 0.49 |   118.47 ± 0.19 |       1.41 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  61 | tg 128     |      4.35 ± 0.03 |     5.01 ± 0.04 |       1.15 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  62 | pp 512     |     85.28 ± 0.23 |   121.99 ± 0.38 |       1.43 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  62 | tg 128     |      4.47 ± 0.05 |     5.67 ± 0.15 |       1.27 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  63 | pp 512     |     86.71 ± 0.23 |   125.13 ± 0.21 |       1.44 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  63 | tg 128     |      4.61 ± 0.03 |     6.14 ± 0.01 |       1.33 |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  64 | pp 512     |     87.99 ± 0.30 |               - |  1.42 (63) |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  64 | tg 128     |      4.74 ± 0.04 |               - | 1.30  (63) |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  65 | pp 512     |     89.19 ± 0.23 |               - | 1.40  (63) |
| 70B mostly Q2_K          |  27.27 GiB |    68.98 B |  65 | tg 128     |      5.00 ± 0.05 |               - | 1.23  (63) |

<details>
<summary>v1</summary>

![e009d7db-6cc2-4353-88cd-9fd84f1eb55f](https://github.com/ggerganov/llama.cpp/assets/2141330/2b3ec198-5b64-4bdb-a8ac-8192a72b7fed)
![9890a361-3b8e-4add-b473-c38d593fe897](https://github.com/ggerganov/llama.cpp/assets/2141330/f270afcc-ee57-4152-b767-07e049bbba5c)
</details>