@ggerganov 2023-10-03
Regardless of the performance effects, this is a good change since it makes the KV cache addressing more intuitive