@oobabooga 2023-11-16
Could this PR, when combined with the performance gains in https://github.com/ggerganov/llama.cpp/pull/3776, allow 70b models in q4_K_M / q4_K_S precision to run on a 3090 at more than 1-2 tokens/second?