@slaren 2023-12-03
Ok, some notes:
- The reason I didn't continue working on this is because of the two copies of `KQ_mask` and `KQ_pos` for CPU and GPU. I was hoping to do this automatically with ggml-backend, but after your graph building refactoring it may be possible to do it cleanly now.
- The KV cache can get quite big, so it should still be possible to choose to not offload the KV cache for low VRAM situations